Реализовано хранение фото помещений на сервере, а в бд только пути к фото. Убраны модели BuildingBindModel и RoomBindModel, теперь их функции взяли стандартные модели. Решён баг с невалидностью Rental и Invoices.

+++ Services/OrganizationService.cs
@@ -64,7 +64,7 @@ namespace RoomRental.Services
                         .Where(r => _context.Buildings
                             .Where(b => organization.OrganizationId == b.OwnerOrganizationId)
                             .Select(b => b.BuildingId)
-                            .Contains(r.BuildingId))
+                            .Contains((int)r.BuildingId))
                         .Select(r => r.RoomId)
                         .ToList();
